requirements.txt
lxml==5.2.2
Nuitka==2.3
ordered-set==4.1.0
pillow==10.3.0
python-pptx==0.6.23
xlrd==2.0.1
XlsxWriter==3.2.0
zstandard==0.22.0
